test: trying to implement first test
