--- test/constants.js
@@ -1,9 +1,0 @@
-'use strict';
-
-exports.KIND_EXIST = 'kind1';
-exports.KIND_NOTEXIST = 'kind2';
-
-exports.ENTRY_ID_EXIST = '000000000000000000000000';
-exports.ENTRY_ID_NOTEXIST = '000000000000000000000001';
-
-exports.NEW_ENTRY_KINDEXIST = { _id: '123' ,field1: 'Hello two'};
