--- test/hds/index.js
@@ -1,3 +1,0 @@
-'use strict';
-
-exports.Kind = require('./kind');
