+++ test/kinds.js
